Cryptanalysis of Substitution cipher
In this assignment the Hill Climbing algorithm is implemented to decrypt the cipher text encrypted using substitution cipher. The scores of the quadgrams in the text are used as a score metric in the hill climbing algorithm (the scores of various quadgrams are written in the 'english_quadgrams.txt' file.